Add custom XLR footprint with bigger holes
Adding a modified version of the "Jack_XLR_Neutrik_NC4MBV_Vertical"
footprint included in the KiCad "Connector_Audio" library.

The NC4MBV footprint included with KiCad uses hole sizes specified by
Neutrik here: https://www.neutrik.com/en/product/nc4mbv-1

Pins 1 through 4 are cylindrical, with a double-taper on the tip
and then a shoulder. The ground pin is rectangular with a shoulder only.

The hole sizes specified by Neutrik are for having the pin shoulders
sit on the PCB. They work perfectly for that - the jack can be
pushed into the PCB with a little force, then the pin shoulders hit the
PCB with a satisfying click, and the jack stays held in the PCB by
friction.

But, I want the full diameter of the pins to go through the holes and
the black plastic housing to sit on the PCB.

Changes to the footprint:

* For pads 1 through 4: increase hole size from 1.6mm to 1.95mm.
    * Lead diameter measured with calipers is 1.75mm.
    * Add 0.20mm clearance per https://klc.kicad.org/footprint/f7/f7.6/.

* For pads 1 and 4 only: increase pad size from 3.4mm to 4mm.
     * The PCB will be supported by the XLR connectors so I want a
       strong solder connection.
     * Can't increase pad size on pads 2 and 3 because they're already
       pretty close to each other.

* For the G pad: change shape from circle to oval.
    * The pin is rectangular. The size measured with calipers is
      2.00mm width and 0.50mm height.
    * Per https://klc.kicad.org/footprint/f7/f7.7/:
        * Set the hole height to be the pin height plus 0.20mm.
          0.5mm + 0.2mm = 0.7mm.
        * Set the hole width to be the pin width plus the hole height.
          2mm + 0.7mm = 2.7mm.
    * Make the pad 2mm larger than the hole for strong solder.

* Add more ref texts above and below the pads on both front and back
  silkscreen. Helpful for keeping track of where you are when soldering
  under a microscope.

* TODO: fix silkscreen arc intersecting the G pad.